+++ paymentsheet-example/src/main/java/com/stripe/android/paymentsheet/example/playground/settings/CurrencySettingsDefinition.kt
@@ -37,6 +37,7 @@ enum class Currency(val displayName: String, override val value: String) : Value
     USD("USD", "usd"),
     INR("INR", "inr"),
     IDR("IDR", "idr"),
+    PHP("PHP", "php"),
     VND("VND", "vnd"),
     PLN("PLN", "pln"),
     SGD("SGD", "sgd"),